requirements.txt
Django==1.11.6
django-countries==4.6.2
django-extensions==1.9.1
PyYAML==3.12
